+++ cmsdev/internal/lib/k8s/k8s.go
@@ -105,6 +105,24 @@ func GetKubectlPath() (string, error) {
 	return KubectlPath, nil
 }
 
+func GetTenants() (tenantList []string, err error) {
+	var path string
+	var cmdList = []string{"get", "tenants", "-n", "tenants", "-o", "custom-columns=:.metadata.name ", "--no-headers"}
+	path, err = GetKubectlPath()
+	if err != nil {
+		return
+	}
+	cmd := exec.Command(path, cmdList...)
+	common.Debugf("Running command: %s", cmd)
+	cmdOut, err := cmd.CombinedOutput()
+	if err != nil {
+		return
+	}
+	// Trim final newline and split by newline
+	tenantList = strings.Split(strings.TrimSpace(string(cmdOut)), "\n")
+	return
+}
+
 func RunCommandInContainer(podName, namespace, containerName string, cmdStrings ...string) (string, error) {
 	k8sCmdList := [...]string{"exec", podName, "-n", namespace, "-c", containerName, "--stdin=false", "--"}
 	cmdList := append(k8sCmdList[:], cmdStrings...)